ll - ApplicationRouter throws exception when the path isn't resolvable
Previously, the code raises an exception for calling a method on a nil
object when a path specified in the AePageObject::Document is not resolvable.

This is confusing to the consumer because:
1. in the stacktrace, it doesn't complain at the line the causes the
problem.
2. the error message itself isn't the root cause of the problem.

To reproduce this issue, refer to the Github issue https://github.com/appfolio/ae_page_objects/issues/163