Remove waiting from ElementProxy methods
The `ElementProxy` class has a set of methods that query the state of
the element.

```
visible?
hidden?
present?
absent?
```

These methods all include waiting, so instead of just querying the state
they *wait* for the condition to be `true`.

This causes considerable performance issues due to developers not being
aware of this behavior. Since we have other methods for waiting for
elements to be visible / hidden / present / absent, these methods should
not wait at all.