Add a textarea element
Add new clear options to set a value for textareas. By default, in
chrome selenium test, when `set` value of text-related node, it would
only prepend the first line. As the capybara chrome node: it only send
a space and a backspace to clear the text. https://bit.ly/3zpJZYJ
However, this will **ONLY** clear the first line of the textarea.

In this commit, we added a Textarea element to override the `set` method
to clear all text in the textarea.

This is following the set definition in the Capybara: it says we could
use `[[:command, 'a'], :backspace]` to clear all text.
Github capybara/selenium/node.rb#L55: https://bit.ly/3CtlTP3

It works both for firefox and chrome